Moving shape and shape_helpers to lowlevel
fix import blocks_list_to_npy

fix : FakeAgent imports SPECIAL_SHAPE_FNS

fix locobot imports
