+++ agents/craftassist/tests/test_memory.py
@@ -5,6 +5,8 @@ Copyright (c) Facebook, Inc. and its affiliates.
 import unittest
 
 import droidlet.base_util
+import droidlet.lowlevel.minecraft.shape_helpers
+import droidlet.lowlevel.minecraft.shapes
 from droidlet.memory.craftassist.mc_memory import MCAgentMemory
 from droidlet.base_util import MOBS_BY_ID
 from droidlet.interpreter.craftassist import dance
@@ -33,8 +35,8 @@ class ObjectsTest(BaseCraftassistTestCase):
 class TriggersTests(BaseCraftassistTestCase):
     def setUp(self):
         super().setUp()
-        self.cube_right = self.add_object(droidlet.base_util.cube(bid=(42, 0)), (9, 63, 4))
-        self.cube_left = self.add_object(droidlet.base_util.cube(), (9, 63, 10))
+        self.cube_right = self.add_object(droidlet.lowlevel.minecraft.shapes.cube(bid=(42, 0)), (9, 63, 4))
+        self.cube_left = self.add_object(droidlet.lowlevel.minecraft.shapes.cube(), (9, 63, 10))
         self.set_looking_at(list(self.cube_right.blocks.keys())[0])
 
     def test_workspace_cleared_on_object_delete(self):
