+++ droidlet/interpreter/craftassist/mc_interpreter.py
@@ -221,7 +221,7 @@ class MCInterpreter(Interpreter):
             ]
         else:  # a schematic
             interprets = interpret_schematic(
-                self, speaker, d.get("schematic", {}), self.block_data
+                self, speaker, d.get("schematic", {}), self.block_data, self.special_shape_functions
             )
 
         # Get the locations to build
@@ -439,7 +439,7 @@ class MCInterpreter(Interpreter):
             if location_d is not None:
                 rd = location_d.get("relative_direction")
                 if rd is not None and (
-                    rd == "AROUND" or rd == "CLOCKWISE" or rd == "ANTICLOCKWISE"
+                        rd == "AROUND" or rd == "CLOCKWISE" or rd == "ANTICLOCKWISE"
                 ):
                     ref_obj = None
                     location_reference_object = location_d.get("reference_object")
