+++ droidlet/lowlevel/minecraft/cuberite_process.py
@@ -12,7 +12,7 @@ import tempfile
 import time
 from droidlet.lowlevel.minecraft.craftassist_cuberite_utils import place_blocks, edit_cuberite_config
 from droidlet.lowlevel.minecraft.craftassist_cuberite_utils.wait_for_cuberite import wait_for_cuberite
-from droidlet.base_util import build_shape_scene
+from droidlet.lowlevel.minecraft.shape_helpers import build_shape_scene
 
 logging.basicConfig(format="%(asctime)s [%(levelname)s]: %(message)s")
 logging.getLogger().setLevel(logging.DEBUG)
