+++ droidlet/lowlevel/minecraft/shapes.py
@@ -7,9 +7,11 @@ their arrangements"""
 import math
 import numpy as np
 from typing import Optional, Tuple
+from typing import Dict, Callable
 
 IDM = Tuple[int, int]
 DEFAULT_IDM = (5, 0)
+
 
 # TODO cylinder
 # TODO: add negative versions of each of the shapes
@@ -21,7 +23,7 @@ DEFAULT_IDM = (5, 0)
 
 
 def hollow_triangle(
-    size=3, depth=1, bid=DEFAULT_IDM, orient="xy", thickness=1, labelme=False, **kwargs
+        size=3, depth=1, bid=DEFAULT_IDM, orient="xy", thickness=1, labelme=False, **kwargs
 ):
     """
     Construct an empty isosceles triangle with a given half base length (b/2).
@@ -58,14 +60,14 @@ def hollow_triangle(
 
 
 def hollow_rectangle(
-    size: Tuple[int, int] = (5, 3),
-    height: Optional[int] = None,
-    length: Optional[int] = None,
-    bid: IDM = DEFAULT_IDM,
-    thickness=1,
-    orient="xy",
-    labelme=False,
-    **kwargs,
+        size: Tuple[int, int] = (5, 3),
+        height: Optional[int] = None,
+        length: Optional[int] = None,
+        bid: IDM = DEFAULT_IDM,
+        thickness=1,
+        orient="xy",
+        labelme=False,
+        **kwargs,
 ):
     R = rectangle(size=size, height=height, length=length, bid=bid, orient=orient)
     l = [r[0] for r in R]
@@ -87,13 +89,13 @@ def hollow_rectangle(
 
 
 def rectangle(
-    size=(5, 3),
-    height: Optional[int] = None,
-    length: Optional[int] = None,
-    bid: IDM = DEFAULT_IDM,
-    orient="xy",
-    labelme=False,
-    **kwargs,
+        size=(5, 3),
+        height: Optional[int] = None,
+        length: Optional[int] = None,
+        bid: IDM = DEFAULT_IDM,
+        orient="xy",
+        labelme=False,
+        **kwargs,
 ):
     if type(size) is int:
         size = [size, size]
@@ -148,7 +150,7 @@ def triangle(size=3, bid=DEFAULT_IDM, orient="xy", thickness=1, labelme=False, *
 
 
 def circle(
-    radius=4, size=None, bid=DEFAULT_IDM, orient="xy", thickness=1, labelme=False, **kwargs
+        radius=4, size=None, bid=DEFAULT_IDM, orient="xy", thickness=1, labelme=False, **kwargs
 ):
     if size is not None:
         radius = size // 2
@@ -208,7 +210,7 @@ def disk(radius=5, size=None, bid=DEFAULT_IDM, orient="xy", thickness=1, labelme
 
 
 def rectanguloid(
-    size=None, depth=None, height=None, width=None, bid=DEFAULT_IDM, labelme=False, **kwargs
+        size=None, depth=None, height=None, width=None, bid=DEFAULT_IDM, labelme=False, **kwargs
 ):
     """Construct a solid rectanguloid"""
     if type(size) is int:
@@ -251,7 +253,7 @@ def near_extremes(x, a, b, r):
 
 
 def rectanguloid_frame(
-    size=3, thickness=1, bid=DEFAULT_IDM, only_corners=False, labelme=False, **kwargs
+        size=3, thickness=1, bid=DEFAULT_IDM, only_corners=False, labelme=False, **kwargs
 ):
     """Construct just the lines of a rectanguloid"""
     R = hollow_rectanguloid(size=size, thickness=thickness, bid=bid, labelme=False, **kwargs)
@@ -383,7 +385,7 @@ def spherical_shell(radius=5, size=None, thickness=2, bid=DEFAULT_IDM, labelme=F
 
 
 def square_pyramid(
-    slope=1, radius=10, size=None, height=None, bid=DEFAULT_IDM, labelme=False, **kwargs
+        slope=1, radius=10, size=None, height=None, bid=DEFAULT_IDM, labelme=False, **kwargs
 ):
     if size is not None:
         radius = size + 2  # this is a heuristic
@@ -607,20 +609,20 @@ def get_rect_instance_seg(bx, by, bz):
     ]
 
     I["top_edge"] = [
-        tuple((s, by[1], bz[i]) for s in range(bx[0], bx[1] + 1)) for i in range(2)
-    ] + [tuple((bx[i], by[1], s) for s in range(bz[0], bz[1] + 1)) for i in range(2)]
+                        tuple((s, by[1], bz[i]) for s in range(bx[0], bx[1] + 1)) for i in range(2)
+                    ] + [tuple((bx[i], by[1], s) for s in range(bz[0], bz[1] + 1)) for i in range(2)]
 
     I["bottom_edge"] = [
-        tuple((s, by[0], bz[i]) for s in range(bx[0], bx[1] + 1)) for i in range(2)
-    ] + [tuple((bx[i], by[0], s) for s in range(bz[0], bz[1] + 1)) for i in range(2)]
+                           tuple((s, by[0], bz[i]) for s in range(bx[0], bx[1] + 1)) for i in range(2)
+                       ] + [tuple((bx[i], by[0], s) for s in range(bz[0], bz[1] + 1)) for i in range(2)]
 
     I["face"] = [
-        tuple((s, t, bz[i]) for s in range(bx[0], bx[1] + 1) for t in range(by[0], by[1] + 1))
-        for i in range(2)
-    ] + [
-        tuple((bx[i], t, s) for s in range(bz[0], bz[1] + 1) for t in range(by[0], by[1] + 1))
-        for i in range(2)
-    ]
+                    tuple((s, t, bz[i]) for s in range(bx[0], bx[1] + 1) for t in range(by[0], by[1] + 1))
+                    for i in range(2)
+                ] + [
+                    tuple((bx[i], t, s) for s in range(bz[0], bz[1] + 1) for t in range(by[0], by[1] + 1))
+                    for i in range(2)
+                ]
 
     I["top"] = [
         tuple((s, by[1], t) for s in range(bx[0], bx[1] + 1) for t in range(bz[0], bz[1] + 1))
@@ -674,38 +676,30 @@ def mirror(S, axis=0):
 # arrangement will eventually be handled by relpos model
 # for now it is either a 'circle' or a 'line'
 # schematic is the thing to be built at each location
-def arrange(arrangement, schematic=None, shapeparams={}):
-    """This function arranges an Optional schematic in a given arrangement
-    and returns the offsets"""
-    N = shapeparams.get("N", 7)
-    extra_space = shapeparams.get("extra_space", 1)
-    if schematic is None:
-        bounds = [0, 1, 0, 1, 0, 1]
-    else:
-        bounds = get_bounds(schematic)
-    if N > 0:
-        if arrangement == "circle":
-            orient = shapeparams.get("orient", "xy")
-            encircled_object_radius = shapeparams.get("encircled_object_radius", 1)
-            b = max(bounds[1] - bounds[0], bounds[3] - bounds[2], bounds[5] - bounds[4])
-            radius = max(((b + extra_space) * N) / (2 * np.pi), encircled_object_radius + b + 1)
-            offsets = [
-                (radius * np.cos(2 * s * np.pi / N), 0, radius * np.sin(2 * s * np.pi / N))
-                for s in range(N)
-            ]
-            if orient == "yz":
-                offsets = [np.round(np.asarray(0, offsets[i][0], offsets[i][2])) for i in range(N)]
-            if orient == "xz":
-                offsets = [
-                    np.round(np.asarray((offsets[i][0], offsets[i][2], 0))) for i in range(N)
-                ]
-        elif arrangement == "line":
-            orient = shapeparams.get("orient")  # this is a vector here
-            b = max(bounds[1] - bounds[0], bounds[3] - bounds[2], bounds[5] - bounds[4])
-            b += extra_space + 1
-            offsets = [np.round(i * b * np.asarray(orient)) for i in range(N)]
-    if N <= 0:
-        raise NotImplementedError(
-            "TODO arrangement just based on extra space, need to specify number for now"
-        )
-    return offsets
+
+
+def cube(size=3, bid=DEFAULT_IDM, labelme=False, **kwargs):
+    if type(size) not in (tuple, list):
+        size = (size, size, size)
+    return rectanguloid(size=size, bid=bid, labelme=labelme)
+
+
+# mapping from canonicalized shape names to the corresponding functions
+SPECIAL_SHAPE_FNS: Dict[str, Callable] = {
+    "CUBE": cube,
+    "HOLLOW_CUBE": hollow_cube,
+    "RECTANGULOID": rectanguloid,
+    "HOLLOW_RECTANGULOID": hollow_rectanguloid,
+    "SPHERE": sphere,
+    "SPHERICAL_SHELL": spherical_shell,
+    "PYRAMID": square_pyramid,
+    "SQUARE": square,
+    "RECTANGLE": rectangle,
+    "CIRCLE": circle,
+    "DISK": disk,
+    "TRIANGLE": triangle,
+    "DOME": dome,
+    "ARCH": arch,
+    "ELLIPSOID": ellipsoid,
+    "TOWER": tower,
+}
